+++ nilearn/plotting/displays/__init__.py
@@ -1,0 +1,28 @@
+"""
+Bl
+"""
+
+from .axes import BaseAxes, CutAxes, GlassBrainAxes
+
+from .projectors import(
+    OrthoProjector, XZProjector, YZProjector,
+    YXProjector, XProjector, YProjector, ZProjector,
+    LZRYProjector, LYRZProjector, LYRProjector,
+    LZRProjector, LRProjector, LProjector, RProjector,
+    get_projector,
+)
+
+from .slicers import(
+    OrthoSlicer, TiledSlicer, MosaicSlicer,
+    XZSlicer, YZSlicer, YXSlicer, XSlicer, YSlicer,
+    ZSlicer, get_slicer,
+)
+
+__all__ = ["BaseAxes", "CutAxes", "GlassBrainAxes",
+           "OrthoProjector", "XZProjector", "YZProjector",
+           "YXProjector", "XProjector", "YProjector", "ZProjector",
+           "LZRYProjector", "LYRZProjector", "LYRProjector",
+           "LZRProjector", "LRProjector", "LProjector", "RProjector",
+           "OrthoSlicer", "TiledSlicer", "MosaicSlicer",
+           "XZSlicer", "YZSlicer", "YXSlicer", "XSlicer", "YSlicer",
+           "ZSlicer", "get_projector", "get_slicer"]
