+++ nilearn/plotting/displays/slicers.py
@@ -1,585 +1,30 @@
-"""
-The Slicer classes.
 
-The main purpose of these classes is to have auto adjust of axes size to
-the data with different layout of cuts.
-"""
-
-import collections.abc
 import numbers
-from distutils.version import LooseVersion
+import collections
+import numpy as np
 
 import matplotlib
-import matplotlib.pyplot as plt
-import numpy as np
-import warnings
 from matplotlib import cm as mpl_cm
-from matplotlib import (colors,
-                        lines,
-                        transforms,
-                        )
+import matplotlib.pyplot as plt
+from matplotlib.colors import LinearSegmentedColormap, ListedColormap
 from matplotlib.colorbar import ColorbarBase
-from matplotlib.font_manager import FontProperties
-from matplotlib.patches import FancyArrow
-from mpl_toolkits.axes_grid1.anchored_artists import AnchoredSizeBar
-from scipy import sparse, stats
+from matplotlib.transforms import Bbox
+
+from distutils.version import LooseVersion
+
+from nilearn._utils import check_niimg_3d
+from nilearn.plotting.find_cuts import find_xyz_cut_coords, find_cut_slices
+from nilearn.plotting.displays import CutAxes
+from nilearn.plotting.edge_detect import _edge_map
+from nilearn.image.resampling import get_bounds, get_mask_bounds
+from nilearn.image import reorder_img, new_img_like, get_data
+from nilearn._utils.niimg import _is_binary_niimg, _safe_get_data
+from nilearn.plotting.displays.axes import _coords_3d_to_2d
 
-from . import cm, glass_brain
-from .edge_detect import _edge_map
-from .find_cuts import find_xyz_cut_coords, find_cut_slices
-from .. import _utils
-from ..image import new_img_like
-from ..image.resampling import (get_bounds, reorder_img, coord_transform,
-                                get_mask_bounds)
-from nilearn.image import get_data
-
-
-###############################################################################
-# class BaseAxes
-###############################################################################
-
-class BaseAxes(object):
-    """ An MPL axis-like object that displays a 2D view of 3D volumes
-    """
-
-    def __init__(self, ax, direction, coord):
-        """ An MPL axis-like object that displays a cut of 3D volumes
-
-        Parameters
-        ----------
-        ax : A MPL axes instance
-            The axes in which the plots will be drawn.
-
-        direction : {'x', 'y', 'z'}
-            The directions of the view.
-
-        coord : float
-            The coordinate along the direction of the cut.
-
-        """
-        self.ax = ax
-        self.direction = direction
-        self.coord = coord
-        self._object_bounds = list()
-        self.shape = None
-
-    def transform_to_2d(self, data, affine):
-        raise NotImplementedError("'transform_to_2d' needs to be implemented "
-                                  "in derived classes'")
-
-    def add_object_bounds(self, bounds):
-        """Ensures that axes get rescaled when adding object bounds
-
-        """
-        old_object_bounds = self.get_object_bounds()
-        self._object_bounds.append(bounds)
-        new_object_bounds = self.get_object_bounds()
-
-        if new_object_bounds != old_object_bounds:
-            self.ax.axis(self.get_object_bounds())
-
-    def draw_2d(self, data_2d, data_bounds, bounding_box,
-                type='imshow', **kwargs):
-        # kwargs messaging
-        kwargs['origin'] = 'upper'
-
-        if self.direction == 'y':
-            (xmin, xmax), (_, _), (zmin, zmax) = data_bounds
-            (xmin_, xmax_), (_, _), (zmin_, zmax_) = bounding_box
-        elif self.direction in 'xlr':
-            (_, _), (xmin, xmax), (zmin, zmax) = data_bounds
-            (_, _), (xmin_, xmax_), (zmin_, zmax_) = bounding_box
-        elif self.direction == 'z':
-            (xmin, xmax), (zmin, zmax), (_, _) = data_bounds
-            (xmin_, xmax_), (zmin_, zmax_), (_, _) = bounding_box
-        else:
-            raise ValueError('Invalid value for direction %s' %
-                             self.direction)
-        ax = self.ax
-        # Here we need to do a copy to avoid having the image changing as
-        # we change the data
-        im = getattr(ax, type)(data_2d.copy(),
-                               extent=(xmin, xmax, zmin, zmax),
-                               **kwargs)
-
-        self.add_object_bounds((xmin_, xmax_, zmin_, zmax_))
-        self.shape = data_2d.T.shape
-
-        # The bounds of the object do not take into account a possible
-        # inversion of the axis. As such, we check that the axis is properly
-        # inverted when direction is left
-        if self.direction == 'l' and not (ax.get_xlim()[0] > ax.get_xlim()[1]):
-            ax.invert_xaxis()
-
-        return im
-
-    def get_object_bounds(self):
-        """ Return the bounds of the objects on this axes.
-        """
-        if len(self._object_bounds) == 0:
-            # Nothing plotted yet
-            return -.01, .01, -.01, .01
-        xmins, xmaxs, ymins, ymaxs = np.array(self._object_bounds).T
-        xmax = max(xmaxs.max(), xmins.max())
-        xmin = min(xmins.min(), xmaxs.min())
-        ymax = max(ymaxs.max(), ymins.max())
-        ymin = min(ymins.min(), ymaxs.min())
-
-        return xmin, xmax, ymin, ymax
-
-    def draw_left_right(self, size, bg_color, **kwargs):
-        if self.direction in 'xlr':
-            return
-        ax = self.ax
-        ax.text(.1, .95, 'L',
-                transform=ax.transAxes,
-                horizontalalignment='left',
-                verticalalignment='top',
-                size=size,
-                bbox=dict(boxstyle="square,pad=0",
-                          ec=bg_color, fc=bg_color, alpha=1),
-                **kwargs)
-
-        ax.text(.9, .95, 'R',
-                transform=ax.transAxes,
-                horizontalalignment='right',
-                verticalalignment='top',
-                size=size,
-                bbox=dict(boxstyle="square,pad=0", ec=bg_color, fc=bg_color),
-                **kwargs)
-
-    def draw_scale_bar(self, bg_color, size=5.0, units='cm',
-                       fontproperties=None, frameon=False, loc=4, pad=.1,
-                       borderpad=.5, sep=5, size_vertical=0, label_top=False,
-                       color='black', fontsize=None, **kwargs):
-        """ Adds a scale bar annotation to the display
-
-        Parameters
-        ----------
-        bgcolor : matplotlib color: str or (r, g, b) value
-            The background color of the scale bar annotation.
-
-        size : float, optional
-            Horizontal length of the scale bar, given in `units`.
-            Default=5.0.
-
-        units : str, optional
-            Physical units of the scale bar (`'cm'` or `'mm'`).
-            Default='cm'.
-
-        fontproperties : ``matplotlib.font_manager.FontProperties`` or dict, optional
-            Font properties for the label text.
-
-        frameon : Boolean, optional
-            Whether the scale bar is plotted with a border. Default=False.
-
-        loc : int, optional
-            Location of this scale bar. Valid location codes are documented
-            `here <https://matplotlib.org/mpl_toolkits/axes_grid/\
-            api/anchored_artists_api.html#mpl_toolkits.axes_grid1.\
-            anchored_artists.AnchoredSizeBar>`__.
-            Default=4.
-
-        pad : int of float, optional
-            Padding around the label and scale bar, in fraction of the font
-            size. Default=0.1.
-
-        borderpad : int or float, optional
-            Border padding, in fraction of the font size. Default=0.5.
-
-        sep : int or float, optional
-            Separation between the label and the scale bar, in points.
-            Default=5.
-
-        size_vertical : int or float, optional
-            Vertical length of the size bar, given in `units`. Default=0.
-
-        label_top : bool, optional
-            If True, the label will be over the scale bar. Default=False.
-
-        color : str, optional
-            Color for the scale bar and label. Default='black'.
-
-        fontsize : int, optional
-            Label font size (overwrites the size passed in through the
-            ``fontproperties`` argument).
-
-        **kwargs :
-            Keyworded arguments to pass to
-            ``matplotlib.offsetbox.AnchoredOffsetbox``.
-
-        """
-        axis = self.ax
-        fontproperties = fontproperties or FontProperties()
-        if fontsize:
-            fontproperties.set_size(fontsize)
-        width_mm = size
-        if units == 'cm':
-            width_mm *= 10
-
-        anchor_size_bar = AnchoredSizeBar(
-            axis.transData,
-            width_mm,
-            '%g%s' % (size, units),
-            fontproperties=fontproperties,
-            frameon=frameon,
-            loc=loc,
-            pad=pad,
-            borderpad=borderpad,
-            sep=sep,
-            size_vertical=size_vertical,
-            label_top=label_top,
-            color=color,
-            **kwargs)
-
-        if frameon:
-            anchor_size_bar.patch.set_facecolor(bg_color)
-            anchor_size_bar.patch.set_edgecolor('none')
-        axis.add_artist(anchor_size_bar)
-
-    def draw_position(self, size, bg_color, **kwargs):
-        raise NotImplementedError("'draw_position' should be implemented "
-                                  "in derived classes")
-
-
-###############################################################################
-# class CutAxes
-###############################################################################
-
-class CutAxes(BaseAxes):
-    """ An MPL axis-like object that displays a cut of 3D volumes
-    """
-    def transform_to_2d(self, data, affine):
-        """ Cut the 3D volume into a 2D slice
-
-        Parameters
-        ----------
-        data : 3D ndarray
-            The 3D volume to cut.
-
-        affine : 4x4 ndarray
-            The affine of the volume.
-
-        """
-        coords = [0, 0, 0]
-        coords['xyz'.index(self.direction)] = self.coord
-        x_map, y_map, z_map = [int(np.round(c)) for c in
-                               coord_transform(coords[0],
-                                               coords[1],
-                                               coords[2],
-                                               np.linalg.inv(affine))]
-        if self.direction == 'y':
-            cut = np.rot90(data[:, y_map, :])
-        elif self.direction == 'x':
-            cut = np.rot90(data[x_map, :, :])
-        elif self.direction == 'z':
-            cut = np.rot90(data[:, :, z_map])
-        else:
-            raise ValueError('Invalid value for direction %s' %
-                             self.direction)
-        return cut
-
-    def draw_position(self, size, bg_color, decimals=False, **kwargs):
-        if decimals:
-            text = '%s=%.{}f'.format(decimals)
-            coord = float(self.coord)
-        else:
-            text = '%s=%i'
-            coord = self.coord
-        ax = self.ax
-        ax.text(0, 0, text % (self.direction, coord),
-                transform=ax.transAxes,
-                horizontalalignment='left',
-                verticalalignment='bottom',
-                size=size,
-                bbox=dict(boxstyle="square,pad=0",
-                          ec=bg_color, fc=bg_color, alpha=1),
-                **kwargs)
-
-
-def _get_index_from_direction(direction):
-    """Returns numerical index from direction
-    """
-    directions = ['x', 'y', 'z']
-    try:
-        # l and r are subcases of x
-        if direction in 'lr':
-            index = 0
-        else:
-            index = directions.index(direction)
-    except ValueError:
-        message = (
-            '{0} is not a valid direction. '
-            "Allowed values are 'l', 'r', 'x', 'y' and 'z'").format(direction)
-        raise ValueError(message)
-    return index
-
-
-def _coords_3d_to_2d(coords_3d, direction, return_direction=False):
-    """Project 3d coordinates into 2d ones given the direction of a cut
-    """
-    index = _get_index_from_direction(direction)
-    dimensions = [0, 1, 2]
-    dimensions.pop(index)
-
-    if return_direction:
-        return coords_3d[:, dimensions], coords_3d[:, index]
-
-    return coords_3d[:, dimensions]
-
-
-###############################################################################
-# class GlassBrainAxes
-###############################################################################
-
-class GlassBrainAxes(BaseAxes):
-    """An MPL axis-like object that displays a 2D projection of 3D
-    volumes with a schematic view of the brain.
-
-    """
-    def __init__(self, ax, direction, coord, plot_abs=True, **kwargs):
-        super(GlassBrainAxes, self).__init__(ax, direction, coord)
-        self._plot_abs = plot_abs
-        if ax is not None:
-            object_bounds = glass_brain.plot_brain_schematics(ax,
-                                                              direction,
-                                                              **kwargs)
-            self.add_object_bounds(object_bounds)
-
-    def transform_to_2d(self, data, affine):
-        """ Returns the maximum of the absolute value of the 3D volume
-        along an axis.
-
-        Parameters
-        ----------
-        data : 3D ndarray
-            The 3D volume.
-
-        affine : 4x4 ndarray
-            The affine of the volume.
-
-        """
-        if self.direction in 'xlr':
-            max_axis = 0
-        else:
-            max_axis = '.yz'.index(self.direction)
-
-        # set unselected brain hemisphere activations to 0
-
-        if self.direction == 'l':
-            x_center, _, _, _ = np.dot(np.linalg.inv(affine),
-                                       np.array([0, 0, 0, 1]))
-            data_selection = data[:int(x_center), :, :]
-        elif self.direction == 'r':
-            x_center, _, _, _ = np.dot(np.linalg.inv(affine),
-                                       np.array([0, 0, 0, 1]))
-            data_selection = data[int(x_center):, :, :]
-        else:
-            data_selection = data
-
-        # We need to make sure data_selection is not empty in the x axis
-        # This should be the case since we expect images in MNI space
-        if data_selection.shape[0] == 0:
-            data_selection = data
-
-        if not self._plot_abs:
-            # get the shape of the array we are projecting to
-            new_shape = list(data.shape)
-            del new_shape[max_axis]
-
-            # generate a 3D indexing array that points to max abs value in the
-            # current projection
-            a1, a2 = np.indices(new_shape)
-            inds = [a1, a2]
-            inds.insert(max_axis, np.abs(data_selection).argmax(axis=max_axis))
-
-            # take the values where the absolute value of the projection
-            # is the highest
-            maximum_intensity_data = data_selection[tuple(inds)]
-        else:
-            maximum_intensity_data = np.abs(data_selection).max(axis=max_axis)
-
-        # This work around can be removed bumping matplotlib > 2.1.0. See #1815
-        # in nilearn for the invention of this work around
-
-        if self.direction == 'l' and data_selection.min() is np.ma.masked and \
-                not (self.ax.get_xlim()[0] > self.ax.get_xlim()[1]):
-            self.ax.invert_xaxis()
-
-        return np.rot90(maximum_intensity_data)
-
-    def draw_position(self, size, bg_color, **kwargs):
-        # It does not make sense to draw crosses for the position of
-        # the cuts since we are taking the max along one axis
-        pass
-
-    def _add_markers(self, marker_coords, marker_color, marker_size, **kwargs):
-        """Plot markers
-
-        In the case of 'l' and 'r' directions (for hemispheric projections),
-        markers in the coordinate x == 0 are included in both hemispheres.
-
-        """
-        marker_coords_2d = _coords_3d_to_2d(marker_coords, self.direction)
-        xdata, ydata = marker_coords_2d.T
-
-        # Allow markers only in their respective hemisphere when appropriate
-        if self.direction in 'lr':
-            if not isinstance(marker_color, str) and \
-                    not isinstance(marker_color, np.ndarray):
-                marker_color = np.asarray(marker_color)
-            relevant_coords = []
-            xcoords, ycoords, zcoords = marker_coords.T
-            for cidx, xc in enumerate(xcoords):
-                if self.direction == 'r' and xc >= 0:
-                    relevant_coords.append(cidx)
-                elif self.direction == 'l' and xc <= 0:
-                    relevant_coords.append(cidx)
-            xdata = xdata[relevant_coords]
-            ydata = ydata[relevant_coords]
-            # if marker_color is string for example 'red' or 'blue', then
-            # we pass marker_color as it is to matplotlib scatter without
-            # making any selection in 'l' or 'r' color.
-            # More likely that user wants to display all nodes to be in
-            # same color.
-            if not isinstance(marker_color, str) and \
-                    len(marker_color) != 1:
-                marker_color = marker_color[relevant_coords]
-
-            if not isinstance(marker_size, numbers.Number):
-                marker_size = np.asarray(marker_size)[relevant_coords]
-
-        defaults = {'marker': 'o',
-                    'zorder': 1000}
-        for k, v in defaults.items():
-            kwargs.setdefault(k, v)
-
-        self.ax.scatter(xdata, ydata, s=marker_size,
-                        c=marker_color, **kwargs)
-
-    def _add_lines(self, line_coords, line_values, cmap,
-                   vmin=None, vmax=None, directed=False, **kwargs):
-        """Plot lines
-
-        Parameters
-        ----------
-        line_coords : list of numpy arrays of shape (2, 3)
-            3d coordinates of lines start points and end points.
-
-        line_values : array_like
-            Values of the lines.
-
-        cmap : colormap
-            Colormap used to map line_values to a color.
-
-        vmin, vmax : float, optional
-            If not None, either or both of these values will be used to
-            as the minimum and maximum values to color lines. If None are
-            supplied the maximum absolute value within the given threshold
-            will be used as minimum (multiplied by -1) and maximum
-            coloring levels.
-
-        directed : boolean, optional
-            Add arrows instead of lines if set to True. Use this when plotting
-            directed graphs for example. Default=False.
-
-        kwargs : dict
-            Additional arguments to pass to matplotlib Line2D.
-
-        """
-        # colormap for colorbar
-        self.cmap = cmap
-        if vmin is None and vmax is None:
-            abs_line_values_max = np.abs(line_values).max()
-            vmin = -abs_line_values_max
-            vmax = abs_line_values_max
-        elif vmin is None:
-            if vmax > 0:
-                vmin = -vmax
-            else:
-                raise ValueError(
-                    "If vmax is set to a non-positive number "
-                    "then vmin needs to be specified"
-                )
-        elif vmax is None:
-            if vmin < 0:
-                vmax = -vmin
-            else:
-                raise ValueError(
-                    "If vmin is set to a non-negative number "
-                    "then vmax needs to be specified"
-                )
-        norm = colors.Normalize(vmin=vmin,
-                                vmax=vmax)
-        # normalization useful for colorbar
-        self.norm = norm
-        abs_norm = colors.Normalize(vmin=0,
-                                    vmax=vmax)
-        value_to_color = plt.cm.ScalarMappable(norm=norm, cmap=cmap).to_rgba
-
-        # Allow lines only in their respective hemisphere when appropriate
-        if self.direction in 'lr':
-            relevant_lines = []
-            for lidx, line in enumerate(line_coords):
-                if self.direction == 'r':
-                    if line[0, 0] >= 0 and line[1, 0] >= 0:
-                        relevant_lines.append(lidx)
-                elif self.direction == 'l':
-                    if line[0, 0] < 0 and line[1, 0] < 0:
-                        relevant_lines.append(lidx)
-            line_coords = np.array(line_coords)[relevant_lines]
-            line_values = line_values[relevant_lines]
-
-        for start_end_point_3d, line_value in zip(
-                line_coords, line_values):
-            start_end_point_2d = _coords_3d_to_2d(start_end_point_3d,
-                                                  self.direction)
-
-            color = value_to_color(line_value)
-            abs_line_value = abs(line_value)
-            linewidth = 1 + 2 * abs_norm(abs_line_value)
-            # Hacky way to put the strongest connections on top of the weakest
-            # note sign does not matter hence using 'abs'
-            zorder = 10 + 10 * abs_norm(abs_line_value)
-            this_kwargs = {'color': color, 'linewidth': linewidth,
-                           'zorder': zorder}
-            # kwargs should have priority over this_kwargs so that the
-            # user can override the default logic
-            this_kwargs.update(kwargs)
-            xdata, ydata = start_end_point_2d.T
-            # If directed is True, add an arrow
-            if directed:
-                dx = xdata[1] - xdata[0]
-                dy = ydata[1] - ydata[0]
-                # Hack to avoid empty arrows to crash with
-                # matplotlib versions older than 3.1
-                # This can be removed once support for
-                # matplotlib pre 3.1 has been dropped.
-                if dx == 0 and dy == 0:
-                    arrow = FancyArrow(xdata[0], ydata[0],
-                                       dx, dy)
-                else:
-                    arrow = FancyArrow(xdata[0], ydata[0],
-                                       dx, dy,
-                                       length_includes_head=True,
-                                       width=linewidth,
-                                       head_width=3*linewidth,
-                                       **this_kwargs)
-                self.ax.add_patch(arrow)
-            # Otherwise a line
-            else:
-                line = lines.Line2D(xdata, ydata, **this_kwargs)
-                self.ax.add_line(line)
-
-
-###############################################################################
-# class BaseSlicer
-###############################################################################
 
 class BaseSlicer(object):
-    """ The main purpose of these class is to have auto adjust of axes size
-        to the data with different layout of cuts.
-
+    """BaseSlicer implementation which main purpose is to auto adjust
+    the axes size to the data with different layout of cuts.
     """
     # This actually encodes the figsize for only one axe
     _default_figsize = [2.2, 2.6]
@@ -641,7 +86,7 @@ class BaseSlicer(object):
         "Initialize the slicer with an image"
         # deal with "fake" 4D images
         if img is not None and img is not False:
-            img = _utils.check_niimg_3d(img)
+            img = check_niimg_3d(img)
 
         cut_coords = cls.find_cut_coords(img, threshold, cut_coords)
 
@@ -768,7 +213,7 @@ class BaseSlicer(object):
         else:
             self._colorbar = colorbar
 
-        img = _utils.check_niimg_3d(img)
+        img = check_niimg_3d(img)
 
         # Make sure that add_overlay shows consistent default behavior
         # with plot_stat_map
@@ -841,14 +286,14 @@ class BaseSlicer(object):
         # case where this image is binary, such as when this function
         # is called from `add_contours`, continuous interpolation
         # does not make sense and we turn to nearest interpolation instead.
-        if _utils.niimg._is_binary_niimg(img):
+        if _is_binary_niimg(img):
             img = reorder_img(img, resample='nearest')
         else:
             img = reorder_img(img, resample=resampling_interpolation)
         threshold = float(threshold) if threshold is not None else None
 
         if threshold is not None:
-            data = _utils.niimg._safe_get_data(img, ensure_finite=True)
+            data = _safe_get_data(img, ensure_finite=True)
             if threshold == 0:
                 data = np.ma.masked_equal(data, 0, copy=False)
             else:
@@ -857,7 +302,7 @@ class BaseSlicer(object):
             img = new_img_like(img, data, img.affine)
 
         affine = img.affine
-        data = _utils.niimg._safe_get_data(img, ensure_finite=True)
+        data = _safe_get_data(img, ensure_finite=True)
         data_bounds = get_bounds(data.shape, affine)
         (xmin, xmax), (ymin, ymax), (zmin, zmax) = data_bounds
 
@@ -969,9 +414,9 @@ class BaseSlicer(object):
         if norm.vmin == norm.vmax:  # len(np.unique(data)) == 1 ?
             return
         else:
-            our_cmap = colors.LinearSegmentedColormap.from_list(
-                'Custom cmap', cmaplist, our_cmap.N)
-
+            our_cmap = LinearSegmentedColormap.from_list(
+                'Custom cmap', cmaplist, our_cmap.N
+            )
         self._cbar = ColorbarBase(
             self._colorbar_ax, ticks=ticks, norm=norm,
             orientation='vertical', cmap=our_cmap, boundaries=bounds,
@@ -1005,7 +450,7 @@ class BaseSlicer(object):
         img = reorder_img(img, resample='continuous')
         data = get_data(img)
         affine = img.affine
-        single_color_cmap = colors.ListedColormap([color])
+        single_color_cmap = ListedColormap([color])
         data_bounds = get_bounds(data.shape, img.affine)
 
         # For each ax, cut the data and plot it
@@ -1314,8 +759,8 @@ class OrthoSlicer(BaseSlicer):
             for prev_ax in direction_ax[:idx]:
                 left_dict[ax] += width_dict[prev_ax]
 
-        return transforms.Bbox([[left_dict[axes], y0],
-                               [left_dict[axes] + width_dict[axes], y1]])
+        return Bbox([[left_dict[axes], y0],
+                     [left_dict[axes] + width_dict[axes], y1]])
 
     def draw_cross(self, cut_coords=None, **kwargs):
         """ Draw a crossbar on the plot to show where the cut is
@@ -1370,10 +815,6 @@ class OrthoSlicer(BaseSlicer):
             if y is not None:
                 ax.axhline(y, **kwargs)
 
-
-###############################################################################
-# class TiledSlicer
-###############################################################################
 
 class TiledSlicer(BaseSlicer):
     """ A class to create 3 axes for plotting orthogonal
@@ -1646,8 +1087,8 @@ class TiledSlicer(BaseSlicer):
                 rel_width_dict, rel_height_dict,
                 rect_x0, rect_y0, rect_x1, rect_y1)
 
-        return transforms.Bbox([[coord1[axes], coord2[axes]],
-                               [coord3[axes], coord4[axes]]])
+        return Bbox([[coord1[axes], coord2[axes]],
+                     [coord3[axes], coord4[axes]]])
 
     def draw_cross(self, cut_coords=None, **kwargs):
         """Draw a crossbar on the plot to show where the cut is performed.
@@ -1702,20 +1143,17 @@ class TiledSlicer(BaseSlicer):
             if y is not None:
                 ax.axhline(y, **kwargs)
 
-###############################################################################
-# class BaseStackedSlicer
-###############################################################################
 
 class BaseStackedSlicer(BaseSlicer):
-    """ A class to create linked axes for plotting stacked
+    """A class to create linked axes for plotting stacked
     cuts of 2D maps.
 
     Attributes
     ----------
-    axes : dictionary of axes
+    axes : :obj:`dict` of :class:`~matplotlib.Axes`
         The axes used to plot each view.
 
-    frame_axes : axes
+    frame_axes : :class:`~matplotlib.Axes`
         The axes framing the whole set of views.
 
     Notes
@@ -1726,7 +1164,7 @@ class BaseStackedSlicer(BaseSlicer):
     """
     @classmethod
     def find_cut_coords(cls, img=None, threshold=None, cut_coords=None):
-        "Instantiate the slicer and find cut coordinates"
+        "Instantiate the slicer and find cut coordinates."
         if cut_coords is None:
             cut_coords = 7
 
@@ -1773,9 +1211,8 @@ class BaseStackedSlicer(BaseSlicer):
             self.frame_axes.set_zorder(-1000)
 
     def _locator(self, axes, renderer):
-        """ The locator function used by matplotlib to position axes.
+        """The locator function used by matplotlib to position axes.
         Here we put the logic used to adjust the size of the axes.
-
         """
         x0, y0, x1, y1 = self.rect
         width_dict = dict()
@@ -1806,16 +1243,16 @@ class BaseStackedSlicer(BaseSlicer):
             left_dict[display_ax.ax] = left
             this_width = width_dict[display_ax.ax]
             left += this_width
-        return transforms.Bbox([[left_dict[axes], y0],
-                                [left_dict[axes] + width_dict[axes], y1]])
+        return Bbox([[left_dict[axes], y0],
+                     [left_dict[axes] + width_dict[axes], y1]])
 
     def draw_cross(self, cut_coords=None, **kwargs):
-        """ Draw a crossbar on the plot to show where the cut is
+        """Draw a crossbar on the plot to show where the cut is
         performed.
 
         Parameters
         ----------
-        cut_coords : 3-tuple of floats, optional
+        cut_coords : 3-:obj:`tuple` of :obj:`float`, optional
             The position of the cross to draw. If none is passed, the
             ortho_slicer's cut coordinates are used.
 
@@ -1827,42 +1264,48 @@ class BaseStackedSlicer(BaseSlicer):
 
 
 class XSlicer(BaseStackedSlicer):
+    """XSlicer class."""
     _direction = 'x'
     _default_figsize = [2.6, 2.3]
 
 
 class YSlicer(BaseStackedSlicer):
+    """YSlicer class."""
     _direction = 'y'
     _default_figsize = [2.2, 2.3]
 
 
 class ZSlicer(BaseStackedSlicer):
+    """ZSlicer class."""
     _direction = 'z'
     _default_figsize = [2.2, 2.3]
 
 
 class XZSlicer(OrthoSlicer):
+    """XZSlicer class."""
     _cut_displayed = 'xz'
 
 
 class YXSlicer(OrthoSlicer):
+    """YXSlicer class."""
     _cut_displayed = 'yx'
 
 
 class YZSlicer(OrthoSlicer):
+    """YZSlicer class."""
     _cut_displayed = 'yz'
 
 
 class MosaicSlicer(BaseSlicer):
-    """ A class to create 3 axes for plotting cuts of 3D maps,
-    in multiple rows and columns.
+    """A class to create 3 :class:`~matplotlib.Axes` for
+    plotting cuts of 3D maps, in multiple rows and columns.
 
     Attributes
     ----------
-    axes : dictionary of axes
+    axes : :obj:`dict` of :class:`~matplotlib.Axes`
         The 3 axes used to plot multiple views.
 
-    frame_axes : axes
+    frame_axes : :class:`~matplotlib.Axes`
         The axes framing the whole set of views.
 
     """
@@ -1876,24 +1319,25 @@ class MosaicSlicer(BaseSlicer):
 
         Parameters
         ----------
-        img : 3D Nifti1Image, optional
+        img : 3D :class:`~nibabel.nifti1.Nifti1Image`, optional
             The brain image.
 
-        threshold : float, optional
-            The lower threshold to the positive activation. If None, the
-            activation threshold is computed using the 80% percentile of
+        threshold : :obj:`float`, optional
+            The lower threshold to the positive activation. If ``None``,
+            the activation threshold is computed using the 80% percentile of
             the absolute value of the map.
 
-        cut_coords : list/tuple of 3 floats, integer, optional
-            xyz world coordinates of cuts. If cut_coords
+        cut_coords : :obj:`list` / :obj:`tuple` of 3 :obj:`float`,\
+        :obj:`int`, optional
+            xyz world coordinates of cuts. If ``cut_coords``
             are not provided, 7 coordinates of cuts are automatically
             calculated.
 
         Returns
         -------
-        cut_coords : dict
-            xyz world coordinates of cuts in a direction. Each key
-            denotes the direction.
+        cut_coords : :obj:`dict`
+            xyz world coordinates of cuts in a direction.
+            Each key denotes the direction.
         """
         if cut_coords is None:
             cut_coords = 7
@@ -1916,25 +1360,27 @@ class MosaicSlicer(BaseSlicer):
 
     @staticmethod
     def _find_cut_coords(img, cut_coords, cut_displayed):
-        """ Find slicing positions along a given axis.
+        """Find slicing positions along a given axis.
 
-            Helper function to find_cut_coords.
+        Helper function to :func:`~nilearn.plotting.find_cut_coords`.
 
         Parameters
         ----------
-        img : 3D Nifti1Image
+        img : 3D :class:`~nibabel.nifti1.Nifti1Image`
             The brain image.
 
-        cut_coords : list/tuple of 3 floats, integer, optional
+        cut_coords : :obj:`list` / :obj:`tuple` of 3 :obj:`float`,\
+        :obj:`int`, optional
             xyz world coordinates of cuts.
 
-        cut_displayed : str
-            Sectional directions 'yxz'
+        cut_displayed : :obj:`str`
+            Sectional directions 'yxz'.
 
         Returns
         -------
-        cut_coords : 1D array of length specified in n_cuts
-            The computed cut_coords.
+        cut_coords : 1D :class:`~np.ndarray` of length specified\
+        in ``n_cuts``
+            The computed ``cut_coords``.
         """
         coords = dict()
         if img is None or img is False:
@@ -1957,7 +1403,7 @@ class MosaicSlicer(BaseSlicer):
         Parameters
         ----------
         kwargs:
-            additional arguments to pass to self._axes_class
+            additional arguments to pass to ``self._axes_class``.
 
         """
         if not isinstance(self.cut_coords, dict):
@@ -1999,8 +1445,8 @@ class MosaicSlicer(BaseSlicer):
                 ax.set_axes_locator(self._locator)
 
     def _locator(self, axes, renderer):
-        """ The locator function used by matplotlib to position axes.
-            Here we put the logic used to adjust the size of the axes.
+        """The locator function used by matplotlib to position axes.
+        Here we put the logic used to adjust the size of the axes.
         """
         x0, y0, x1, y1 = self.rect
         display_ax_dict = self.axes
@@ -2046,9 +1492,8 @@ class MosaicSlicer(BaseSlicer):
                     left += this_width
                     bottom_dict[display_ax.ax] = fraction * index * (y1 - y0)
                     height_dict[display_ax.ax] = this_height
-        return transforms.Bbox([[left_dict[axes], bottom_dict[axes]],
-                                [left_dict[axes] + width_dict[axes],
-                                 height_dict[axes]]])
+        return Bbox([[left_dict[axes], bottom_dict[axes]],
+                     [left_dict[axes] + width_dict[axes], height_dict[axes]]])
 
 
     def draw_cross(self, cut_coords=None, **kwargs):
@@ -2077,269 +1522,12 @@ SLICERS = dict(ortho=OrthoSlicer,
                z=ZSlicer)
 
 
-class OrthoProjector(OrthoSlicer):
-    """A class to create linked axes for plotting orthogonal projections
-    of 3D maps.
-
-    """
-    _axes_class = GlassBrainAxes
-
-    @classmethod
-    def find_cut_coords(cls, img=None, threshold=None, cut_coords=None):
-        return (None, ) * len(cls._cut_displayed)
-
-    def draw_cross(self, cut_coords=None, **kwargs):
-        # It does not make sense to draw crosses for the position of
-        # the cuts since we are taking the max along one axis
-        pass
-
-    def add_graph(self, adjacency_matrix, node_coords,
-                  node_color='auto', node_size=50,
-                  edge_cmap=cm.bwr,
-                  edge_vmin=None, edge_vmax=None,
-                  edge_threshold=None,
-                  edge_kwargs=None, node_kwargs=None, colorbar=False,
-                  ):
-        """Plot undirected graph on each of the axes
-
-        Parameters
-        ----------
-        adjacency_matrix : numpy array of shape (n, n)
-            Represents the edges strengths of the graph.
-            The matrix can be symmetric which will result in
-            an undirected graph, or not symmetric which will
-            result in a directed graph.
-
-        node_coords : numpy array_like of shape (n, 3)
-            3d coordinates of the graph nodes in world space.
-
-        node_color : color or sequence of colors, optional
-            Color(s) of the nodes. Default='auto'.
-
-        node_size : scalar or array_like, optional
-            Size(s) of the nodes in points^2. Default=50.
-
-        edge_cmap : colormap, optional
-            Colormap used for representing the strength of the edges.
-            Default=cm.bwr.
-
-        edge_vmin, edge_vmax : float, optional
-            If not None, either or both of these values will be used to
-            as the minimum and maximum values to color edges. If None are
-            supplied the maximum absolute value within the given threshold
-            will be used as minimum (multiplied by -1) and maximum
-            coloring levels.
-
-        edge_threshold : str or number, optional
-            If it is a number only the edges with a value greater than
-            edge_threshold will be shown.
-            If it is a string it must finish with a percent sign,
-            e.g. "25.3%", and only the edges with a abs(value) above
-            the given percentile will be shown.
-
-        edge_kwargs : dict, optional
-            Will be passed as kwargs for each edge matlotlib Line2D.
-
-        node_kwargs : dict
-            Will be passed as kwargs to the plt.scatter call that plots all
-            the nodes in one go.
-
-        """
-        # set defaults
-        if edge_kwargs is None:
-            edge_kwargs = {}
-        if node_kwargs is None:
-            node_kwargs = {}
-        if isinstance(node_color, str) and node_color == 'auto':
-            nb_nodes = len(node_coords)
-            node_color = mpl_cm.Set2(np.linspace(0, 1, nb_nodes))
-        node_coords = np.asarray(node_coords)
-
-        # decompress input matrix if sparse
-        if sparse.issparse(adjacency_matrix):
-            adjacency_matrix = adjacency_matrix.toarray()
-
-        # make the lines below well-behaved
-        adjacency_matrix = np.nan_to_num(adjacency_matrix)
-
-        # safety checks
-        if 's' in node_kwargs:
-            raise ValueError("Please use 'node_size' and not 'node_kwargs' "
-                             "to specify node sizes")
-        if 'c' in node_kwargs:
-            raise ValueError("Please use 'node_color' and not 'node_kwargs' "
-                             "to specify node colors")
-
-        adjacency_matrix_shape = adjacency_matrix.shape
-        if (len(adjacency_matrix_shape) != 2 or
-                adjacency_matrix_shape[0] != adjacency_matrix_shape[1]):
-            raise ValueError(
-                "'adjacency_matrix' is supposed to have shape (n, n)."
-                ' Its shape was {0}'.format(adjacency_matrix_shape))
-
-        node_coords_shape = node_coords.shape
-        if len(node_coords_shape) != 2 or node_coords_shape[1] != 3:
-            message = (
-                "Invalid shape for 'node_coords'. You passed an "
-                "'adjacency_matrix' of shape {0} therefore "
-                "'node_coords' should be a array with shape ({0[0]}, 3) "
-                'while its shape was {1}').format(adjacency_matrix_shape,
-                                                  node_coords_shape)
-
-            raise ValueError(message)
-
-        if isinstance(node_color, (list, np.ndarray)) and len(node_color) != 1:
-            if len(node_color) != node_coords_shape[0]:
-                raise ValueError(
-                    "Mismatch between the number of nodes ({0}) "
-                    "and the number of node colors ({1})."
-                    .format(node_coords_shape[0], len(node_color)))
-
-        if node_coords_shape[0] != adjacency_matrix_shape[0]:
-            raise ValueError(
-                "Shape mismatch between 'adjacency_matrix' "
-                "and 'node_coords'"
-                "'adjacency_matrix' shape is {0}, 'node_coords' shape is {1}"
-                .format(adjacency_matrix_shape, node_coords_shape))
-
-        # If the adjacency matrix is not symmetric, give a warning
-        symmetric = True
-        if not np.allclose(adjacency_matrix, adjacency_matrix.T, rtol=1e-3):
-            symmetric = False
-            warnings.warn(("'adjacency_matrix' is not symmetric. "
-                           "A directed graph will be plotted."))
-
-        # For a masked array, masked values are replaced with zeros
-        if hasattr(adjacency_matrix, 'mask'):
-            if not (adjacency_matrix.mask == adjacency_matrix.mask.T).all():
-                symmetric = False
-                warnings.warn(("'adjacency_matrix' was masked with "
-                               "a non symmetric mask. A directed "
-                               "graph will be plotted."))
-            adjacency_matrix = adjacency_matrix.filled(0)
-
-        if edge_threshold is not None:
-            if symmetric:
-                # Keep a percentile of edges with the highest absolute
-                # values, so only need to look at the covariance
-                # coefficients below the diagonal
-                lower_diagonal_indices = np.tril_indices_from(adjacency_matrix,
-                                                              k=-1)
-                lower_diagonal_values = adjacency_matrix[
-                    lower_diagonal_indices]
-                edge_threshold = _utils.param_validation.check_threshold(
-                    edge_threshold, np.abs(lower_diagonal_values),
-                    stats.scoreatpercentile, 'edge_threshold')
-            else:
-                edge_threshold = _utils.param_validation.check_threshold(
-                    edge_threshold, np.abs(adjacency_matrix.ravel()),
-                    stats.scoreatpercentile, 'edge_threshold')
-
-            adjacency_matrix = adjacency_matrix.copy()
-            threshold_mask = np.abs(adjacency_matrix) < edge_threshold
-            adjacency_matrix[threshold_mask] = 0
-
-        if symmetric:
-            lower_triangular_adjacency_matrix = np.tril(adjacency_matrix, k=-1)
-            non_zero_indices = lower_triangular_adjacency_matrix.nonzero()
-        else:
-            non_zero_indices = adjacency_matrix.nonzero()
-
-        line_coords = [node_coords[list(index)]
-                       for index in zip(*non_zero_indices)]
-
-        adjacency_matrix_values = adjacency_matrix[non_zero_indices]
-        for ax in self.axes.values():
-            ax._add_markers(node_coords, node_color, node_size, **node_kwargs)
-            if line_coords:
-                ax._add_lines(line_coords, adjacency_matrix_values, edge_cmap,
-                              vmin=edge_vmin, vmax=edge_vmax, directed=(not symmetric),
-                              **edge_kwargs)
-            # To obtain the brain left view, we simply invert the x axis
-            if ax.direction == 'l' and not (ax.ax.get_xlim()[0] > ax.ax.get_xlim()[1]):
-                ax.ax.invert_xaxis()
-
-        if colorbar:
-            self._colorbar = colorbar
-            self._show_colorbar(ax.cmap, ax.norm, threshold=edge_threshold)
-
-        plt.draw_if_interactive()
+def get_slicer(display_mode):
+    "Internal function to retrieve a slicer"
+    return _get_create_display_fun(display_mode, SLICERS)
 
 
-class XProjector(OrthoProjector):
-    _cut_displayed = 'x'
-    _default_figsize = [2.6, 2.3]
-
-
-class YProjector(OrthoProjector):
-    _cut_displayed = 'y'
-    _default_figsize = [2.2, 2.3]
-
-
-class ZProjector(OrthoProjector):
-    _cut_displayed = 'z'
-    _default_figsize = [2.2, 2.3]
-
-
-class XZProjector(OrthoProjector):
-    _cut_displayed = 'xz'
-
-
-class YXProjector(OrthoProjector):
-    _cut_displayed = 'yx'
-
-
-class YZProjector(OrthoProjector):
-    _cut_displayed = 'yz'
-
-
-class LYRZProjector(OrthoProjector):
-    _cut_displayed = 'lyrz'
-
-
-class LZRYProjector(OrthoProjector):
-    _cut_displayed = 'lzry'
-
-
-class LZRProjector(OrthoProjector):
-    _cut_displayed = 'lzr'
-
-
-class LYRProjector(OrthoProjector):
-    _cut_displayed = 'lyr'
-
-
-class LRProjector(OrthoProjector):
-    _cut_displayed = 'lr'
-
-
-class LProjector(OrthoProjector):
-    _cut_displayed = 'l'
-    _default_figsize = [2.6, 2.3]
-
-
-class RProjector(OrthoProjector):
-    _cut_displayed = 'r'
-    _default_figsize = [2.6, 2.3]
-
-
-PROJECTORS = dict(ortho=OrthoProjector,
-                  xz=XZProjector,
-                  yz=YZProjector,
-                  yx=YXProjector,
-                  x=XProjector,
-                  y=YProjector,
-                  z=ZProjector,
-                  lzry=LZRYProjector,
-                  lyrz=LYRZProjector,
-                  lyr=LYRProjector,
-                  lzr=LZRProjector,
-                  lr=LRProjector,
-                  l=LProjector,
-                  r=RProjector)
-
-
-def get_create_display_fun(display_mode, class_dict):
+def _get_create_display_fun(display_mode, class_dict):
     try:
         return class_dict[display_mode].init_with_figure
     except KeyError:
