+++ nilearn/plotting/img_plotting.py
@@ -42,7 +42,7 @@ from ..image import new_img_like, iter_img, get_data, math_img, resample_to_img
 from ..input_data import NiftiMasker
 from nilearn.image.resampling import reorder_img
 from ..masking import compute_epi_mask, apply_mask
-from .displays import get_slicer, get_projector
+from nilearn.plotting.displays import get_slicer, get_projector
 from . import cm
